after first deployment

+++ src/components/myTeams.jsx
@@ -3,6 +3,8 @@ import PageHeader from "./common/pageHeader";
 import teamService from "../services/teamService";
 import Team from "./team";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
+import { toast } from "react-toastify";
 
 class MyTeams extends Component {
   state = {
@@ -15,10 +17,23 @@ class MyTeams extends Component {
   }
 
   removeTeam = (teamId) => {
-    let { teams } = this.state;
-    teams = teams.filter((team) => team._id !== teamId);
-    teamService.deleteTeam(teamId);
-    this.setState({ teams });
+    Swal.fire({
+      title: "Are you sure you want to delete this team?",
+      text: "you wont be able to cancel",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        let { teams } = this.state;
+        teams = teams.filter((team) => team._id !== teamId);
+        teamService.deleteTeam(teamId);
+        this.setState({ teams });
+        toast("This team was deleted succesfully");
+      }
+    });
   };
   render() {
     const { teams } = this.state;
@@ -33,8 +48,12 @@ class MyTeams extends Component {
             <Link className="btn btn-primary ms-2" to="/join-team">
               Join An Existing Team
             </Link>
+            <p>
+              You can try a test team.
+              <br /> enter "Join an existing team" and try out
+            </p>
             <p className="mt-3 team-list">
-              {teams.length > 0 && <p>Your teams in the list below:</p>}
+              {teams.length > 0 && <p>Your teams are in the list below:</p>}
             </p>
           </div>
         </div>
